Move OffchainRanomBalancing to common
