+++ runtime/westend/src/lib.rs
@@ -347,33 +347,6 @@ parameter_types! {
 	pub OffchainRepeat: BlockNumber = 5;
 }
 
-/// Maximum number of iterations for balancing that will be executed in the embedded miner of
-/// the pallet.
-pub const MINER_MAX_ITERATIONS: u32 = 10;
-
-/// A source of random balance for the NPoS Solver, which is meant to be run by the OCW election
-// miner.
-pub struct OffchainRandomBalance;
-impl frame_support::pallet_prelude::Get<Option<(usize, sp_npos_elections::ExtendedBalance)>>
-	for OffchainRandomBalance
-{
-	fn get() -> Option<(usize, sp_npos_elections::ExtendedBalance)> {
-		use sp_runtime::traits::TrailingZeroInput;
-		let iters = match MINER_MAX_ITERATIONS {
-			0 => 0,
-			max @ _ => {
-				let seed = sp_io::offchain::random_seed();
-				let random = <u32>::decode(&mut TrailingZeroInput::new(&seed))
-					.expect("input is padded with zeroes; qed") %
-					max.saturating_add(1);
-				random as usize
-			},
-		};
-
-		Some((iters, 0))
-	}
-}
-
 sp_npos_elections::generate_solution_type!(
 	#[compact]
 	pub struct NposCompactSolution16::<
@@ -409,7 +382,7 @@ impl pallet_election_provider_multi_phase::Config for Runtime {
 	type Solver = frame_election_provider_support::SequentialPhragmen<
 		AccountId,
 		pallet_election_provider_multi_phase::SolutionAccuracyOf<Runtime>,
-		OffchainRandomBalance,
+		runtime_common::elections::OffchainRandomBalancing,
 	>;
 	type BenchmarkingConfig = runtime_common::elections::BenchmarkConfig;
 	type ForceOrigin = EnsureRoot<AccountId>;
